// src/models/_foodb-models/db.d.ts
export interface healthEffectsAttribute {
	id: number;
	name: string;
	description?: string;
	chebiName?: string;
	chebiId?: string;
	createdAt?: Date;
	updatedAt?: Date;
	creatorId?: number;
	updaterId?: number;
	chebiDefinition?: string;
}

export interface healthEffectsInstance extends Sequelize.Instance<healthEffectsAttribute>, healthEffectsAttribute {
}